+++ app/Http/Resources/PaginateResource.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Http\Resources;
+
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\ResourceCollection;
+
+class PaginateResource extends ResourceCollection
+{
+    /**
+     * Transform the resource into an array.
+     *
+     * @return array<string, mixed>
+     */
+    public function toArray(Request $request) : array
+    {
+        return [
+            'per_page' => $this->perPage(),
+            'current_page' => $this->currentPage(),
+            'total' => $this->total(),
+            'last_page' => $this->lastPage(),
+            'from' => $this->firstItem(),
+            'to' => $this->lastItem()
+        ];
+    }
+}
